tensorflow_test/requirements.txt
absl-py==0.8.0
astor==0.8.0
certifi==2020.4.5.1
cycler==0.10.0
decorator==4.4.1
gast==0.3.2
grpcio==1.16.1
h5py==2.9.0
imageio==2.6.1
joblib==0.14.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
networkx==2.4
nltk==3.4.5
numpy==1.16.5
pandas==0.25.1
Pillow==6.2.1
playsound==1.2.2
protobuf==3.9.2
PyAudio==0.2.11
pyparsing==2.4.2
pyreadline==2.1
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tornado==6.0.3
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
